adjustments to make this not crash

+++ database/migrations/2022_09_07_174413_common_countries.php
@@ -14,6 +14,8 @@ class CommonCountries extends Migration
     public function up()
     {
 
+        if(!Schema::hasColumn('geo_countries', 'is_common')) {
+            
             Schema::table('geo_countries', function (Blueprint $table) {
                 $table->tinyInteger('is_common')->default(0)->index();
             });
@@ -21,6 +23,8 @@ class CommonCountries extends Migration
             $uk = \AscentCreative\Geo\Models\Country::find('1188');
             $uk->is_common = true;
             $uk->save();        
+
+        }
 
     }
 
@@ -31,8 +35,10 @@ class CommonCountries extends Migration
      */
     public function down()
     {
-        Schema::table('geo_countries', function (Blueprint $table) {
-            $table->dropColumn('is_common');
-        });
+        if(Schema::hasColumn('geo_countries', 'is_common')) {
+            Schema::table('geo_countries', function (Blueprint $table) {
+                $table->dropColumn('is_common');
+            });
+        }
     }
 }
